+++ app/Http/Controllers/Admin/Gostaresh/AssetProductivityController.php
@@ -6,6 +6,8 @@ use App\Exports\Gostaresh\AssetProductivity\ListExport;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Gostaresh\AssetProductivity\AssetProductivityRequest;
 use App\Models\Index\IndexOfAssetProductivity;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Auth;
 use Maatwebsite\Excel\Facades\Excel;
 use PDF;
@@ -16,10 +18,12 @@ class AssetProductivityController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function index()
     {
+        $this->authorize("view-any-IndexOfAssetProductivity");
+
         $query = IndexOfAssetProductivity::whereRequestsQuery();
 
         $filterColumnsCheckBoxes = IndexOfAssetProductivity::$filterColumnsCheckBoxes;
@@ -34,11 +38,12 @@ class AssetProductivityController extends Controller
             , 'yearSelectedList', 'filterColumnsCheckBoxes'
         ));
     }
+
     private function yearSelectedList($query)
     {
         return $query->select('year')->distinct()->pluck('year');
     }
-    
+
     // ****************** Export ******************
     private function getAssetProductivityRecords()
     {
@@ -64,14 +69,16 @@ class AssetProductivityController extends Controller
         return view('admin.gostaresh.asset-productivity.list.pdf', compact('assetProductivity'));
     }
     // ****************** End Export ******************
-    
+
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function create()
     {
+        $this->authorize("create-any-IndexOfAssetProductivity");
+
         return view('admin.gostaresh.asset-productivity.create.create');
     }
 
@@ -79,11 +86,12 @@ class AssetProductivityController extends Controller
      * Store a newly created resource in storage.
      *
      * @param AssetProductivityRequest $request
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function store(AssetProductivityRequest $request)
     {
-         IndexOfAssetProductivity::create(array_merge(['user_id' => Auth::id()], $request->validated()));
+        $this->authorize("create-any-IndexOfAssetProductivity");
+        IndexOfAssetProductivity::create(array_merge(['user_id' => Auth::id()], $request->validated()));
         return redirect()->back()->with('success', __('titles.success_store'));
     }
 
@@ -102,10 +110,12 @@ class AssetProductivityController extends Controller
      * Show the form for editing the specified resource.
      *
      * @param IndexOfAssetProductivity $assetProductivity
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function edit(IndexOfAssetProductivity $assetProductivity)
     {
+        $this->authorize("edit-any-IndexOfAssetProductivity");
+
         return view('admin.gostaresh.asset-productivity.edit.edit', compact('assetProductivity'));
     }
 
@@ -114,10 +124,12 @@ class AssetProductivityController extends Controller
      *
      * @param AssetProductivityRequest $request
      * @param IndexOfAssetProductivity $assetProductivity
-     * @return \Illuminate\Http\Response
+     * @return Response
+     * @throws AuthorizationException
      */
     public function update(AssetProductivityRequest $request, IndexOfAssetProductivity $assetProductivity)
     {
+        $this->authorize("edit-any-IndexOfAssetProductivity");
         $assetProductivity->update($request->validated());
         return back()->with('success', __('titles.success_update'));
     }
@@ -126,10 +138,11 @@ class AssetProductivityController extends Controller
      * Remove the specified resource from storage.
      *
      * @param IndexOfAssetProductivity $assetProductivity
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function destroy(IndexOfAssetProductivity $assetProductivity)
     {
+        $this->authorize("delete-any-IndexOfAssetProductivity");
         $assetProductivity->delete();
         return back()->with('success', __('titles.success_delete'));
     }
