+++ app/Http/Controllers/Admin/Gostaresh/AverageTestScoreOfTheLastFivePercentOfAdmittedController.php
@@ -4,10 +4,10 @@ namespace App\Http\Controllers\Admin\Gostaresh;
 
 use App\Exports\Gostaresh\AverageTestScoreOfTheLastFivePercentOfAdmitted\ListExport;
 use App\Http\Controllers\Controller;
-use App\Models\Index\AverageTestScoreOfTheLastFivePercentOfAdmitted;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\Gostaresh\AverageTestScoreOfTheLastFivePercentOfAdmitted\AverageTestScoreOfTheLastFivePercentOfAdmittedRequest;
+use App\Models\Index\AverageTestScoreOfTheLastFivePercentOfAdmitted;
+use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
 use Maatwebsite\Excel\Facades\Excel;
 use PDF;
 
@@ -17,10 +17,12 @@ class AverageTestScoreOfTheLastFivePercentOfAdmittedController extends Controlle
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function index()
     {
+        $this->authorize("view-any-AverageTestScoreOfTheLastFivePercentOfAdmitted");
+
         $query = AverageTestScoreOfTheLastFivePercentOfAdmitted::whereRequestsQuery();
 
         $filterColumnsCheckBoxes = AverageTestScoreOfTheLastFivePercentOfAdmitted::$filterColumnsCheckBoxes;
@@ -36,7 +38,7 @@ class AverageTestScoreOfTheLastFivePercentOfAdmittedController extends Controlle
     {
         return $query->select('year')->distinct()->pluck('year');
     }
-    
+
     private function getAverageTestScoreOfTheLastFivePercentOfAdmittedRecords()
     {
         return AverageTestScoreOfTheLastFivePercentOfAdmitted::whereRequestsQuery()->orderBy('id', 'desc')->get();
@@ -64,21 +66,24 @@ class AverageTestScoreOfTheLastFivePercentOfAdmittedController extends Controlle
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function create()
     {
+        $this->authorize("create-any-AverageTestScoreOfTheLastFivePercentOfAdmitted");
+
         return view('admin.gostaresh.average-test-score-of-the-last-five-percent-of-admitted.create.create');
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  AverageTestScoreOfTheLastFivePercentOfAdmittedRequest  $request
-     * @return \Illuminate\Http\Response
+     * @param AverageTestScoreOfTheLastFivePercentOfAdmittedRequest $request
+     * @return Response
      */
     public function store(AverageTestScoreOfTheLastFivePercentOfAdmittedRequest $request)
     {
+        $this->authorize("create-any-AverageTestScoreOfTheLastFivePercentOfAdmitted");
         AverageTestScoreOfTheLastFivePercentOfAdmitted::create(array_merge(['user_id' => Auth::id()], $request->validated()));
         return back()->with('success', __('titles.success_store'));
     }
@@ -86,8 +91,8 @@ class AverageTestScoreOfTheLastFivePercentOfAdmittedController extends Controlle
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param int $id
+     * @return Response
      */
     public function show(AverageTestScoreOfTheLastFivePercentOfAdmitted $avgTstScOfLastFivePctOfAdmitted)
     {
@@ -97,23 +102,27 @@ class AverageTestScoreOfTheLastFivePercentOfAdmittedController extends Controlle
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param int $id
+     * @return Response
      */
     public function edit(AverageTestScoreOfTheLastFivePercentOfAdmitted $avgTstScOfLastFivePctOfAdmitted)
     {
-        return view('admin.gostaresh.average-test-score-of-the-last-five-percent-of-admitted.edit.edit', compact('avgTstScOfLastFivePctOfAdmitted'));
+        $this->authorize("edit-any-AverageTestScoreOfTheLastFivePercentOfAdmitted");
+
+        return view(
+            'admin.gostaresh.average-test-score-of-the-last-five-percent-of-admitted.edit.edit', compact('avgTstScOfLastFivePctOfAdmitted'));
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  AverageTestScoreOfTheLastFivePercentOfAdmittedRequest  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param AverageTestScoreOfTheLastFivePercentOfAdmittedRequest $request
+     * @param int $id
+     * @return Response
      */
     public function update(AverageTestScoreOfTheLastFivePercentOfAdmittedRequest $request, AverageTestScoreOfTheLastFivePercentOfAdmitted $avgTstScOfLastFivePctOfAdmitted)
     {
+        $this->authorize("edit-any-AverageTestScoreOfTheLastFivePercentOfAdmitted");
         $avgTstScOfLastFivePctOfAdmitted->update($request->validated());
         return back()->with('success', __('titles.success_update'));
     }
@@ -121,11 +130,12 @@ class AverageTestScoreOfTheLastFivePercentOfAdmittedController extends Controlle
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param int $id
+     * @return Response
      */
     public function destroy(AverageTestScoreOfTheLastFivePercentOfAdmitted $avgTstScOfLastFivePctOfAdmitted)
     {
+        $this->authorize("delete-any-AverageTestScoreOfTheLastFivePercentOfAdmitted");
         $avgTstScOfLastFivePctOfAdmitted->delete();
         return back()->with('success', __('titles.success_delete'));
     }
